Refactored the step 3 form, and fixed the issue with submitting
